bug fix + output file name change
fixed a bug where the program was expecting the output folder to be a file when checking and would then panic.

Changed the name of the output images from a timestamp to their coordinates, zoom and max iteration to be able to reproduce the same results multiple times from the file name.